+++ version-info.js
@@ -1,9 +1,10 @@
 window.HrrraVersionInfo = Object.freeze({
-  versionCode: 8,
-  versionName: "1.0.7",
+  versionCode: 9,
+  versionName: "1.0.8",
   whatsNew: [
-  "Added a new badges system.",
-  "It will still be tuned further.",
-  "And next time, there will finally be some sound too!"
-]
+    "First sound tests are here.",
+    "For now they are limited to Level 1 and still incomplete there too.",
+    "This is a fully experimental version, so I expect bugs.",
+    "Expect the next version around Easter."
+  ]
 });